Message cleanup, WorkerBatch compartmentalization
Ob socket close, pending messages are flushed with an error so that they
become requeued.

Previously, WorkerBatch was created as a global variable. Functions
which depend on it have been made into member functions of WorkerBatch,
and the initialization for testing has been moved to only occur when
server.py is run directly.